fixed issues for categories update and delete.improved movie functions

+++ db/db_category.py
@@ -1,6 +1,6 @@
 from .models import Category
 from sqlalchemy.orm import Session
-
+from schemas import CategoryBase, CategoryDisplay
 
 # Create Category
 def add_category(db: Session, category_name: str):
@@ -18,17 +18,22 @@ def get_all_categories(db: Session):
 def get_category(db: Session, category_id: int):
     return db.query(Category).filter(Category.id == category_id).first()
 
+
 # Update Category
-def update_category(db: Session, category_id: int, category_name: str):
-    category = db.query(Category).filter(Category.id == category_id).first()
-    category.category_name = category_name
-    db.commit()
-    db.refresh(category)
+def update_category(db: Session, category_id: int, request: CategoryBase):
+    category = get_category(db, category_id)
+    if category:
+        category.category_name = request.category_name
+        db.commit()
+        db.refresh(category)
     return category
 
 # Delete Category
 def delete_category(db: Session, category_id: int):
-    category = db.query(Category).filter(Category.id == category_id).first()
-    db.delete(category)
-    db.commit()
-    return "Category with id: {category_id} deleted successfully"
+    category = get_category(db, category_id)
+    if category:
+        db.delete(category)
+        db.commit()
+        return True
+    return False
+
